Queries can include periods
If your using a hypersearch component to search for things that contain
periods, like email addresses, as soon as you type the period things
break. Since it's using the query itself as the key, Embers get/set
gets all sorts of crankypants. This commit replaces periods in queries
as dashes.